requirements.txt
# 企业流程架构设计智能体 - 依赖清单

# Web框架
streamlit>=1.28.0

# LLM API
openai>=1.0.0

# 环境变量
python-dotenv>=1.0.0

# Word文档处理
python-docx>=0.8.11

# PDF处理（可选）
PyPDF2>=3.0.0
